editors - register text editors in one place from `contrib`
